Update example_automatic_report.py
Changes I made include adding plots, data, and variables for the current, with data being read at the start, and the relevant plots and variables placed in Chapter 5.2. 

Additionally, I implemented functionality that allows the user to choose which chapters to include in the final document. For example, the user can type 'yes' to include wind data, controlled by the condition 'if wind == True.' I also modified how tables and chapters are numbered; instead of manually writing the numbers, they are now based on variables (e.g., a for chapter number, b for subsection number, fignr for figure number, and tablenr for table number). 

Lastly, I resolved the page numbering issue by moving the page number code, which generates the '1 of 73' format, from the beginning to the end. This change prevents it from being affected by plot-related page shifts.